+++ lib/protos/icon/component.rb
@@ -2,14 +2,7 @@
 
 module Protos
   module Icon
-    class Component < Phlex::HTML
-      def initialize(filepath)
-        @filepath = filepath
-      end
-
-      def view_template
-        unsafe_raw File.read(@filepath)
-      end
+    class Component < Protos::Component
     end
   end
 end