+++ lib/protos/icon/heroicon_component.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Protos
+  module Icon
+    class HeroiconComponent < Component
+      option :variant, reader: false, default: -> { :solid }
+
+      def view_template
+        public_send(@variant)
+      end
+    end
+  end
+end